pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SIM7600g-H_Tutorial"
version = "0.1.0"
description = "Tutorial for SIM7600G-H: Receive and log SMS on Windows"
authors = [{name = "You"}]
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sim7600 = ["py.typed"]